refactor CutOut and RandomErase to inherit from same base class
